fix: skip Vapor-affected analyzers in Vapor/serverless environments
Vapor removes .env, .env.example, webpack.mix.js, and frontend lock files
from deployments. Six analyzers that check for these files now skip on
Vapor/serverless to prevent false positives.

Extracts a shared DetectsDeploymentPlatform trait (refactored from the
inline implementation in DirectoryWritePermissionsAnalyzer) and applies
it to EnvFileAnalyzer, EnvExampleAnalyzer, EnvVariableAnalyzer,
EnvFileSecurityAnalyzer, FrontendVulnerableDependencyAnalyzer, and
MinificationAnalyzer. Adds 14 new tests covering vapor and serverless
skip behavior for all affected analyzers.
