+++ tests/Unit/Analyzers/Security/EnvFileSecurityAnalyzerTest.php
@@ -936,4 +936,27 @@ ENV;
             );
         }
     }
+
+    // =========================================================================
+    // Vapor / Serverless Skip Tests
+    // =========================================================================
+
+    public function test_skips_on_vapor(): void
+    {
+        /** @var \ShieldCI\Analyzers\Security\EnvFileSecurityAnalyzer $analyzer */
+        $analyzer = $this->createAnalyzer();
+        $analyzer->setDeploymentPlatform('vapor');
+
+        $this->assertFalse($analyzer->shouldRun());
+        $this->assertStringContainsString('Vapor', $analyzer->getSkipReason());
+    }
+
+    public function test_skips_on_serverless(): void
+    {
+        /** @var \ShieldCI\Analyzers\Security\EnvFileSecurityAnalyzer $analyzer */
+        $analyzer = $this->createAnalyzer();
+        $analyzer->setDeploymentPlatform('serverless');
+
+        $this->assertFalse($analyzer->shouldRun());
+    }
 }
